Ajuste no CPF e Telefone

+++ ProjetoFinal/src/DAO/UsuarioDAO.java
@@ -23,8 +23,8 @@ public class UsuarioDAO {
 		pstm.setString(1, usuario.getNome());
 		pstm.setString(2, usuario.getEmail());
 		pstm.setString(3, usuario.getEndereço());
-		pstm.setInt(4, usuario.getCpf());
-		pstm.setInt(5, usuario.getTelefone());
+		pstm.setLong(4, usuario.getCpf());
+		pstm.setLong(5, usuario.getTelefone());
 		pstm.setInt(6, usuario.getSenha());
 		
 		pstm.execute();
